Moved Utils WJ LocalAppData filewatching to lazy system
